splash screen left

+++ app/src/main/java/com/app/fruit2you/data/database/entities/FruitItem.kt
@@ -7,7 +7,7 @@ import androidx.room.PrimaryKey
 data class FruitItem(
     @PrimaryKey
     var name: String,
-    var amount: Double,
+    var amount: Int,
     var quantity: Int) {
 
 }
